requirements.txt
# Core
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.66.0

# Web/API
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.31.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
chromadb>=0.4.0

# AI & ML
# Note: Torch CPU version is recommended for Docker builds to save space/time
torch>=2.0.0
sentence-transformers>=2.2.0
groq>=0.4.0

# LlamaIndex
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0

# Parsing/Crawling
beautifulsoup4>=4.12.0
html2text>=2020.1.16
PyMuPDF>=1.23.0
tiktoken>=0.5.0
filelock>=3.13.0
typing-extensions>=4.9.0
httpx>=0.20.0
regex>=2023.0.0
packaging>=23.0